+++ docs/Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.3
+* Added kpref subitems
+* Added DSL markers
+* Added transition utils and other utils
+* Added custom searchview with binders
+
 ## v1.2
 * Fix title attribute in changelog
 * Update support libs